+++ maven_plugin/src/main/java/io/github/chains_project/maven_lockfile/checksum/FileSystemChecksumCalculator.java
@@ -182,18 +182,4 @@ public class FileSystemChecksumCalculator extends AbstractChecksumCalculator {
         return getResolvedFieldInternal(resolveDependency(artifact, pluginBuildingRequest), pluginBuildingRequest)
                 .orElse(ResolvedUrl.of(""));
     }
-
-    @Override
-    public String calculatePomChecksum(Path path) {
-        try {
-            MessageDigest messageDigest = MessageDigest.getInstance(checksumAlgorithm);
-            byte[] fileBuffer = Files.readAllBytes(path);
-            byte[] artifactHash = messageDigest.digest(fileBuffer);
-            BaseEncoding baseEncoding = BaseEncoding.base16();
-            return baseEncoding.encode(artifactHash).toLowerCase(Locale.ROOT);
-        } catch (Exception e) {
-            LOGGER.warn("Could not calculate checksum for pom " + path, e);
-            return "";
-        }
-    }
 }
