+++ maven_plugin/src/main/java/io/github/chains_project/maven_lockfile/checksum/RemoteChecksumCalculator.java
@@ -204,18 +204,4 @@ public class RemoteChecksumCalculator extends AbstractChecksumCalculator {
     public ResolvedUrl getPluginResolvedField(Artifact artifact) {
         return getResolvedFieldInternal(artifact, pluginBuildingRequest).orElse(ResolvedUrl.Unresolved());
     }
-
-    @Override
-    public String calculatePomChecksum(Path path) {
-        try {
-            MessageDigest messageDigest = MessageDigest.getInstance(checksumAlgorithm);
-            byte[] fileBuffer = Files.readAllBytes(path);
-            byte[] artifactHash = messageDigest.digest(fileBuffer);
-            BaseEncoding baseEncoding = BaseEncoding.base16();
-            return baseEncoding.encode(artifactHash).toLowerCase(Locale.ROOT);
-        } catch (Exception e) {
-            LOGGER.warn("Could not calculate checksum for pom " + path, e);
-            return "";
-        }
-    }
 }
